% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_yahoo_url.R
\name{build_yahoo_url}
\alias{build_yahoo_url}
\title{construct URLs required to scrape relevant data from Yahoo Finance}
\usage{
build_yahoo_url(
  tickers,
  type = c("price", "IS", "BS", "CFS"),
  start_date = NULL,
  end_date = NULL
)
}
\arguments{
\item{tickers}{character vector of stock tickers following Yahoo Finance format i.e. JSE tickers have .JO suffix}

\item{type}{character vector indicating which types of data you require.}

\item{start_date}{start date for price data retrieval. Default NULL assumes no price data.}

\item{end_date}{end date for price data retrieval. Default NULL assumes current date.}
}
\value{
tbl_df with cols ticker, page and url
}
\description{
construct URLs required to scrape relevant data from Yahoo Finance
}
