% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_invcom_data.R
\name{clean_invcom_price}
\alias{clean_invcom_price}
\title{clean price data table from Investing.com}
\usage{
clean_invcom_price(scraped_price_data, frequency)
}
\arguments{
\item{scraped_price_data}{tbl_df of scraped data}

\item{frequency}{string indicating the frequency to return price data, e.g. 'monthly'}
}
\value{
object of class \code{tbl_df} with date, variable and value
}
\description{
clean price data table from Investing.com
}
