% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_yahoo_data.R
\name{clean_yahoo_data}
\alias{clean_yahoo_data}
\title{clean Yahoo data}
\usage{
clean_yahoo_data(scraped_data, type, frequency = NULL)
}
\arguments{
\item{scraped_data}{tbl_df of scraped data}

\item{type}{string indicating the data type, e.g. 'IS'}

\item{frequency}{string}
}
\value{
tbl_df
}
\description{
clean Yahoo data
}
