% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_yahoo_data.R
\name{clean_yahoo_fs}
\alias{clean_yahoo_fs}
\title{clean financial statement tables from Yahoo}
\usage{
clean_yahoo_fs(scraped_fs_data, type)
}
\arguments{
\item{scraped_fs_data}{tbl_df of scraped financial statement data}

\item{type}{string indicating the data type, e.g. 'IS'}
}
\value{
object of class \code{tbl_df} with date, variable and value
}
\description{
clean financial statement tables from Yahoo
}
