% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_company_data.R
\name{get_company_data_single}
\alias{get_company_data_single}
\title{Get company data from Yahoo Finance or Investing.com}
\usage{
get_company_data_single(tickers, type)
}
\arguments{
\item{tickers}{character vector of tickers where data is prefixed with the source, e.g. YAH-EXAMPLETICKER}

\item{type}{string specifying the type of data required - one of is, bs, cfs or price}
}
\value{
a data.frame
}
\description{
Get company data from Yahoo Finance or Investing.com
}
