% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_invcom_data.R
\name{get_invcom_data}
\alias{get_invcom_data}
\title{Get Investing.com data}
\usage{
get_invcom_data(
  tickers,
  type = c("price", "IS", "BS", "CFS"),
  start_date = NULL,
  end_date = NULL,
  frequency = NULL
)
}
\arguments{
\item{tickers}{character vector of investing.com tickers}

\item{type}{character vector specifying the type of data required - price, income statement, balance sheet and/or cash flow statement}

\item{start_date}{date}

\item{end_date}{date}

\item{frequency}{string indicating the frequency of data to return - required for price data}
}
\value{
tbl_df with cols ticker, type, scraped_data and clean_data
}
\description{
Get Investing.com data
}
