% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_invcom_data.R
\name{get_invcom_data_list}
\alias{get_invcom_data_list}
\title{scrape Investing.com Finance data}
\usage{
get_invcom_data_list(pjs_session, ticker_tbl, start_date, end_date)
}
\arguments{
\item{pjs_session}{phantom.js session}

\item{ticker_tbl}{tbl_df with ticker, page and url}

\item{start_date}{start date for price data retrieval}

\item{end_date}{end date for price data retrieval}
}
\value{
object of class \code{list} named by each url specific to a ticker-type combination containing scraped data
}
\description{
scrape Investing.com Finance data
}
