% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_invcom_meta.R
\name{get_invcom_meta_data}
\alias{get_invcom_meta_data}
\title{get ticker meta data from investing.com}
\usage{
get_invcom_meta_data(pjs_session)
}
\arguments{
\item{pjs_session}{phantom.js session}
}
\value{
tbl_df
}
\description{
get ticker meta data from investing.com
}
