% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_invcom_price_data.R
\name{get_invcom_price_data}
\alias{get_invcom_price_data}
\title{get historical price data from investing.com}
\usage{
get_invcom_price_data(pjs_session, start_dt, end_dt)
}
\arguments{
\item{pjs_session}{phantom.js session}

\item{start_dt}{start date}

\item{end_dt}{end date}
}
\value{
tbl_df
}
\description{
get historical price data from investing.com
}
