% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ticker_id.R
\name{get_ticker_id}
\alias{get_ticker_id}
\title{Get Ticker ID
Use investing.com to retrieve name info for a given security}
\usage{
get_ticker_id(ticker_id)
}
\arguments{
\item{ticker_id}{string containing the name, ticker or ISIN code of the security of interest.
Best practice is to use the ISIN as that will always return a unique result for the expected security.
Using tickers or names may result in multiple matches in which case the first match is returned.}
}
\value{
data.frame with columns ticker, name and isin
}
\description{
Get Ticker ID
Use investing.com to retrieve name info for a given security
}
