% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yahoo_data.R
\name{get_yahoo_data}
\alias{get_yahoo_data}
\title{get tidy yahoo data}
\usage{
get_yahoo_data(
  tickers,
  type = c("price", "IS", "BS", "CFS"),
  start_date = NULL,
  end_date = NULL,
  frequency = NULL
)
}
\arguments{
\item{tickers}{character vector of stock tickers following Yahoo Finance format i.e. JSE ticker EXAMPLE.JO}

\item{type}{character vector specifying the type of company data you want to retrieve}

\item{start_date}{start date for price data retrieval. Default NULL assumes no price data.}

\item{end_date}{end date for price data retrieval. Default NULL assumes no price data.}

\item{frequency}{string indicating the frequency of data to retrieve (if retrieving price data)}
}
\value{
object of class \code{tbl_df} with ticker, type, nested data.frame of scraped data and nested data.frame of tidy scraped data
}
\description{
get tidy yahoo data
}
