% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yahoo_fs.R
\name{get_yahoo_fs_data}
\alias{get_yahoo_fs_data}
\title{get financial statement data from Yahoo}
\usage{
get_yahoo_fs_data(pjs_session)
}
\arguments{
\item{pjs_session}{phantom.js session}
}
\value{
single column tbl of raw scraped text data
}
\description{
get financial statement data from Yahoo
}
