% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yahoo_meta.R
\name{get_yahoo_meta_data}
\alias{get_yahoo_meta_data}
\title{Get meta data for a ticker from Yahoo Finance}
\usage{
get_yahoo_meta_data(pjs_session, url)
}
\arguments{
\item{pjs_session}{phantom js session}

\item{url}{string indicating the url for a given tickers key statistics}
}
\value{

}
\description{
Get meta data for a ticker from Yahoo Finance
}
