% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yahoo_price.R
\name{get_yahoo_price_data}
\alias{get_yahoo_price_data}
\title{get price data from Yahoo}
\usage{
get_yahoo_price_data(url)
}
\arguments{
\item{url}{string specifying the download link for the price data}
}
\value{
tbl_df with columns, Date, Open, High, Low, Close, Adj.Close, Volume
}
\description{
get price data from Yahoo
}
