% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_invcom_data.R
\name{navigate_invcom_ticker}
\alias{navigate_invcom_ticker}
\title{navigate to ticker page in investing.com}
\usage{
navigate_invcom_ticker(pjs_session, tkr)
}
\arguments{
\item{pjs_session}{phantom.js session}

\item{tkr}{string}
}
\value{

}
\description{
navigate to ticker page in investing.com
}
