% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_invcom_data.R
\name{navigate_ticker_home}
\alias{navigate_ticker_home}
\title{Navigate to the ticker home page}
\usage{
navigate_ticker_home(pjs_session, ticker)
}
\arguments{
\item{pjs_session}{phantom.js session}

\item{ticker}{string indicating the ticker}
}
\description{
Navigate to the ticker home page
}
