% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_invcom_fs.R
\name{pvt_get_invcom_fs_data}
\alias{pvt_get_invcom_fs_data}
\title{get financial statement data from investing.com}
\usage{
pvt_get_invcom_fs_data(pjs_session, type)
}
\arguments{
\item{pjs_session}{phantom.js session}

\item{type}{string indicating the financial statement required. one of 'IS', 'BS' or 'CFS'}
}
\value{
tbl_df
}
\description{
get financial statement data from investing.com
}
