% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_invcom_data.R
\name{pvt_get_meta}
\alias{pvt_get_meta}
\title{private function to navigate home and return the ticker's meta data}
\usage{
pvt_get_meta(pjs_session, ticker)
}
\arguments{
\item{pjs_session}{phantom.js session}

\item{ticker}{string}
}
\value{
data.frame of meta data associated with the ticker
}
\description{
private function to navigate home and return the ticker's meta data
}
