% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yahoo_meta.R
\name{pvt_get_yahoo_meta_data}
\alias{pvt_get_yahoo_meta_data}
\title{Get meta data for a ticker from Yahoo Finance}
\usage{
pvt_get_yahoo_meta_data(pjs_session)
}
\arguments{
\item{pjs_session}{a phantom js session}

\item{url}{string indicating the url for a given tickers key statistics}
}
\value{
tbl_df with columns variable and value, containing the records Name, Market, Currency, Market Cap, Shares Outstanding and Beta
}
\description{
Get meta data for a ticker from Yahoo Finance
}
