% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_invcom_data.R
\name{clean_invcom_data}
\alias{clean_invcom_data}
\title{clean Investing.com data}
\usage{
clean_invcom_data(scraped_data, type, ...)
}
\arguments{
\item{scraped_data}{tbl_df of scraped data}

\item{type}{string indicating the data type, e.g. 'IS'}

\item{...}{arguments to other methods}
}
\value{
tbl_df
}
\description{
clean Investing.com data
}
