% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_invcom_data.R
\name{get_invcom_data}
\alias{get_invcom_data}
\title{Get Investing.com data}
\usage{
get_invcom_data(
  tickers,
  type = c("is", "bs", "cf", "price"),
  start_date = NULL,
  end_date = NULL,
  frequency = NULL
)
}
\arguments{
\item{tickers}{character vector}

\item{type}{character vector}

\item{start_date}{date}

\item{end_date}{date}

\item{frequency}{string}
}
\value{
tbl_df
}
\description{
Get Investing.com data
}
