% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_invcom_meta.R
\name{get_invcom_meta_data}
\alias{get_invcom_meta_data}
\title{Get meta data for security from investing.com}
\usage{
get_invcom_meta_data(pjs_session)
}
\arguments{
\item{pjs_session}{phantom.js session}
}
\value{
data.frame with columns variable of class character and value of class character
}
\description{
Get meta data for security from investing.com
}
