% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yahoo_price.R
\name{get_yahoo_price_data}
\alias{get_yahoo_price_data}
\title{get price data from Yahoo}
\usage{
get_yahoo_price_data(pjs_session, url)
}
\arguments{
\item{pjs_session}{phantom.js session}

\item{url}{string specifying the url for the price data and its relevant parameters}
}
\value{
single column tbl of raw scraped text data
}
\description{
get price data from Yahoo
}
