% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_invcom_data.R
\name{navigate_ticker_home}
\alias{navigate_ticker_home}
\title{Navigate to the ticker home page from the investing.com home page}
\usage{
navigate_ticker_home(pjs_session, home_url, ticker)
}
\arguments{
\item{pjs_session}{phantom.js session}

\item{home_url}{investing.com home url}

\item{ticker}{string indicating the ticker}
}
\description{
Navigate to the ticker home page from the investing.com home page
}
